// src/types/embed.ts
import type { OptionsBase } from "../providers/types/optionsBase.js";
import type { ErrorAISuite } from "./handleErrorResponse.js";
import type { ResultBase } from "./resultBase.js";

export interface EmbeddingRequest {
  /**
   * The text or array of texts to be embedded.
   * Each content must not exceed the maximum token limit for the model.
   */
  content: string | Array<string> | Array<number> | Array<Array<number>>;
  /**
   * The format of the embedding vector.
   * "float" returns the embedding as an array of floats, while "base64" returns the embedding as a base64-encoded string.\
   * If not specified, defaults to "float".
   */
  encodingFormat?: "float" | "base64";
  /**
   * The number of dimensions for the embedding vector.
   * This is an optional parameter that can be used to specify the desired dimensionality of the embedding.
   * If not specified, the model's default dimensionality will be used.
   */
  dimensions?: number;
}

export interface EmbeddingResponse extends ResultBase {
  /**
   * The list of embeddings generated by the model.
   */
  content: Array<Array<number>>;

  /**
   * The name of the model used to generate the embedding.
   */
  model: string;

  /**
   * The object type, which is always "list".
   */
  object: "list";

  /**
   * The usage information for the request.
   */
  usage?: {
    input_tokens: number;
    output_tokens: 0;
    total_tokens: number;
  };

  metadata?: Record<string, unknown>;
}

export type SuccessEmbedding = EmbeddingResponse & {
  success: true;
};

export type ResultEmbedding = SuccessEmbedding | ErrorAISuite;

export interface EmbeddingOptions extends OptionsBase {}